refactor: update React profiler imports and move session ownership to utils
